Extract Image Dimensions
A function to extract the dimensions of an input image by analyzing the pixel data matrix (np array). Also, as the output format expects 3-dimensional output, the function accounts for 1-D vectors and 2-D matrices by initially setting all 3 dimensions to 1. Hence even 1-D vectors are registered as a single plane on the third dimensions, and having width 1 on the second dimension.